Added code for including member diskstore uuid into the metrics namespace of every member
